Parses all WoS field related to authors and makes author table

Relevant fields that are parsed (as in Web of Science Field Tags 2018-06-27):

AU

AF (full names)

C1 adresses

RP reprint address (one you contact for reprint copy)

EM emails

RI researcher ID

OI ORCID Identifier (Open Researcher and Contributor ID)

disambr_make_wos_tsv_authors(
  wos_data_table,
  list_of_author_fields = c("author_order", "author_short_name", "author_initials",
    "author_last_name", "author_full_name", "author_first_names",
    "author_first_full_name", "author_email", "author_researcher_id", "author_orcid",
    "author_affiliations"),
  ...
)

Arguments

wos_data_table

WoS tsv export data.table

list_of_author_fields

list of data to parse to construct (if not all selected as by default it will save time)

...

set verbose if needed here

Value

author table

Details

Fields that are not parsed

BA book - not parsed

BF book - not parsed

CA gp group author (usually organization or group name) - not parsed

BE editors - not parsed