Make a statistics for a collection of sets (specific algorithm) and writes to the file in sets_dir

Make a statistics for a collection of sets (specific algorithm) and writes to the file in sets_dir

disambr_stats(
  sets = NULL,
  sets_dir = getOption("disambr_save_set_dir"),
  name = sets_dir,
  remove_data = TRUE,
  save_rds = TRUE
)

Arguments

sets

Collection of sets

sets_dir

If first arg is not provided search for collection of sets in this directory. The results are saved to this directory as well.

name

Name of the collection of sets

remove_data

whether to remove first 4 sets (assumed to be a data) when sets arg is set

save_rds

save_rds