Filters list of sets

Filters list of sets

disambr_subsets(
  sets_list,
  attribute_value_list,
  which_to_return = c("all", "first", "last"),
  negate_subsets = FALSE
)

Arguments

sets_list

list of sets

attribute_value_list

list of attribute values where list elements name correspond attribute names used for filtering sets

which_to_return

whether to return "all", "first" or "last" set from filtered sets

negate_subsets

whether to return sets that was not matched insted

Value

list of sets or set if which_to_return is ethier "first" or "last"